:art:refactor: move db method to service
